Created GetClosedRangeIncrement to ensure consistency when using these increments.
Renamed the type and parameter of functionType toIntegrate to integrandFn and integrand. This should be clearer than previously.

Removed the unused deltaB from the FluxCalculator class.

Changed the loop in CalculateFlux to be more amenable to parallelism by allocating the entire array first and then doing assignment on a specified